Introduce fully lexically scoped variables
Adds a new notion of block scoped variables, declared using new syntax
"let x =".
Adds new block scope with "begin" and "end" keywords.
Existing block structures like while/if-elseif-else/try/for are also
lexically scopable.
Also adds explicit global variable assignment / declaration via
"global"
By default variables are global, to keep backwards compatibility with
LambdaMOO.

+++ crates/kernel/src/vm/mod.rs
@@ -102,7 +102,7 @@ pub enum ExecutionResult {
     DispatchFork(Fork),
     /// Request dispatch of a builtin function with the given arguments.
     ContinueBuiltin {
-        bf_func_num: usize,
+        bf_func_num: u16,
         arguments: Vec<Var>,
     },
     /// Request that this task be suspended for a duration of time.
